added casting (UINT32) to remove warning messages.
git-svn-id: svn://svn.code.sf.net/p/aceos/code@98 78f6c469-4e3a-0410-bca2-dfd793a85426

+++ src/kernel/i386/debug/gdb.c
@@ -41,7 +41,7 @@ void exceptionHandler(int exc, void *addr)
 		exc != 13 &&    /* GPF (normal in V86 mode) */
 		exc != 8 )     /* double fault (so abnormal we just crash) */
 	{
-		SetIdtGate(exc, addr);
+		SetIdtGate(exc, (UINT32)addr);
 	}
 }
